split session screen preview out of main file
