+++ android/mobile/ui/session/src/main/java/fr/shiningcat/simplehiit/android/mobile/ui/session/SessionScreen.kt
@@ -5,8 +5,6 @@ import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Surface
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -16,29 +14,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.tooling.preview.PreviewFontScale
-import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import androidx.compose.ui.tooling.preview.PreviewScreenSizes
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.window.core.layout.WindowHeightSizeClass
-import androidx.window.core.layout.WindowWidthSizeClass
 import fr.shiningcat.simplehiit.android.mobile.ui.common.UiArrangement
 import fr.shiningcat.simplehiit.android.mobile.ui.common.components.NavigateUpTopBar
-import fr.shiningcat.simplehiit.android.mobile.ui.common.theme.SimpleHiitMobileTheme
 import fr.shiningcat.simplehiit.android.mobile.ui.session.components.SessionSideBarComponent
 import fr.shiningcat.simplehiit.android.mobile.ui.session.contents.SessionContentHolder
 import fr.shiningcat.simplehiit.commonresources.R
 import fr.shiningcat.simplehiit.commonutils.HiitLogger
-import fr.shiningcat.simplehiit.domain.common.models.Exercise
-import fr.shiningcat.simplehiit.domain.common.models.ExerciseSide
-import fr.shiningcat.simplehiit.domain.common.models.SessionStepDisplay
 
 @Composable
 fun SessionScreen(
@@ -92,7 +79,7 @@ fun SessionScreen(
 }
 
 @Composable
-private fun SessionScreen(
+fun SessionScreen(
     navigateUp: () -> Boolean = { true },
     onAbortSession: () -> Unit = {},
     pause: () -> Unit = {},
@@ -188,137 +175,4 @@ private fun SessionScreen(
             )
         }
     }
-}
-
-// Previews
-@PreviewLightDark
-@PreviewFontScale
-@PreviewScreenSizes
-@Composable
-private fun SessionScreenPreview(
-    @PreviewParameter(SessionScreenPreviewParameterProvider::class) viewState: SessionViewState,
-) {
-    val windowSizeClass = currentWindowAdaptiveInfo().windowSizeClass
-    val previewUiArrangement: UiArrangement =
-        if (windowSizeClass.windowWidthSizeClass == WindowWidthSizeClass.EXPANDED) { // typically, a tablet or bigger in landscape
-            UiArrangement.HORIZONTAL
-        } else { // WindowWidthSizeClass.Medium, WindowWidthSizeClass.Compact :
-            if (windowSizeClass.windowHeightSizeClass == WindowHeightSizeClass.COMPACT) { // typically, a phone in landscape
-                UiArrangement.HORIZONTAL
-            } else {
-                UiArrangement.VERTICAL // typically, a phone or tablet in portrait
-            }
-        }
-    SimpleHiitMobileTheme {
-        Surface {
-            SessionScreen(
-                dialogViewState = SessionDialog.None,
-                screenViewState = viewState,
-                uiArrangement = previewUiArrangement,
-            )
-        }
-    }
-}
-
-internal class SessionScreenPreviewParameterProvider : PreviewParameterProvider<SessionViewState> {
-    override val values: Sequence<SessionViewState>
-        get() =
-            sequenceOf(
-                SessionViewState.Loading,
-                SessionViewState.Error("Blabla error code"),
-                SessionViewState.InitialCountDownSession(
-                    countDown =
-                        CountDown(
-                            secondsDisplay = "3",
-                            progress = .5f,
-                            playBeep = true,
-                        ),
-                ),
-                SessionViewState.RunningNominal(
-                    periodType = RunningSessionStepType.REST,
-                    displayedExercise = Exercise.CatBackLegLift,
-                    side = ExerciseSide.RIGHT,
-                    stepRemainingTime = "25s",
-                    stepRemainingPercentage = .53f,
-                    sessionRemainingTime = "16mn 23s",
-                    sessionRemainingPercentage = .24f,
-                ),
-                SessionViewState.RunningNominal(
-                    periodType = RunningSessionStepType.REST,
-                    displayedExercise = Exercise.CatBackLegLift,
-                    side = ExerciseSide.RIGHT,
-                    stepRemainingTime = "25s",
-                    stepRemainingPercentage = .53f,
-                    sessionRemainingTime = "16mn 23s",
-                    sessionRemainingPercentage = .24f,
-                    countDown =
-                        CountDown(
-                            secondsDisplay = "3",
-                            progress = .5f,
-                            playBeep = true,
-                        ),
-                ),
-                SessionViewState.RunningNominal(
-                    periodType = RunningSessionStepType.REST,
-                    displayedExercise = Exercise.CatBackLegLift,
-                    side = ExerciseSide.RIGHT,
-                    stepRemainingTime = "25s",
-                    stepRemainingPercentage = .23f,
-                    sessionRemainingTime = "16mn 23s",
-                    sessionRemainingPercentage = .57f,
-                ),
-                SessionViewState.RunningNominal(
-                    periodType = RunningSessionStepType.WORK,
-                    displayedExercise = Exercise.CrabAdvancedBridge,
-                    side = ExerciseSide.NONE,
-                    stepRemainingTime = "3s",
-                    stepRemainingPercentage = .7f,
-                    sessionRemainingTime = "5mn 12s",
-                    sessionRemainingPercentage = .3f,
-                ),
-                SessionViewState.RunningNominal(
-                    periodType = RunningSessionStepType.WORK,
-                    displayedExercise = Exercise.CrabAdvancedBridge,
-                    side = ExerciseSide.NONE,
-                    stepRemainingTime = "3s",
-                    stepRemainingPercentage = .7f,
-                    sessionRemainingTime = "5mn 12s",
-                    sessionRemainingPercentage = .3f,
-                    countDown =
-                        CountDown(
-                            secondsDisplay = "5",
-                            progress = 0f,
-                            playBeep = false,
-                        ),
-                ),
-                SessionViewState.RunningNominal(
-                    periodType = RunningSessionStepType.WORK,
-                    displayedExercise = Exercise.CrabAdvancedBridge,
-                    side = ExerciseSide.LEFT,
-                    stepRemainingTime = "3s",
-                    stepRemainingPercentage = .5f,
-                    sessionRemainingTime = "5mn 12s",
-                    sessionRemainingPercentage = .2f,
-                ),
-                SessionViewState.Finished(
-                    "16mn",
-                    workingStepsDone =
-                        listOf(
-                            SessionStepDisplay(Exercise.CatBackLegLift, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.CatKneePushUp, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.LungesArmsCrossSide, ExerciseSide.LEFT),
-                            SessionStepDisplay(Exercise.LungesArmsCrossSide, ExerciseSide.RIGHT),
-                            SessionStepDisplay(Exercise.LungesTwist, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.LyingStarToeTouchSitUp, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.LyingSupermanTwist, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.StandingMountainClimber, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.PlankMountainClimber, ExerciseSide.LEFT),
-                            SessionStepDisplay(Exercise.PlankMountainClimber, ExerciseSide.RIGHT),
-                            SessionStepDisplay(Exercise.StandingKickCrunches, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.SquatBasic, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.PlankShoulderTap, ExerciseSide.NONE),
-                            SessionStepDisplay(Exercise.PlankBirdDogs, ExerciseSide.NONE),
-                        ),
-                ),
-            )
 }
